+++ srcs/compass/render_ring.c
@@ -12,99 +12,7 @@
 
 #include "compass.h"
 
-void	drop_the_blur(t_win *win, t_compass *comp, t_hori_line *draw)
-{
-	t_comp_blur	*blur;
-	int			blur_y;
-	int			blur_x;
-
-	if (!comp->blur_on)
-		return ;
-	blur = &comp->blur;
-	blur_y = comp->radius + draw->y - comp->centre.y;
-	blur_x = comp->radius + draw->min_x - comp->centre.x;
-	while (draw->min_x < draw->max_x)
-	{
-		blur_x = comp->radius + draw->min_x - comp->centre.x;
-		win->set_pixel(win, draw->min_x++, draw->y, \
-			blur->verti_blur[blur_y * blur->blur_height + blur_x]);
-	}
-}
-
-void	place_circle_aa(t_win *win, int x, int y, t_render_circ *ren)
-{
-	win->set_pixel(win, ren->centre.x + x, ren->centre.y + y, \
-	gamma_average(win->get_pixel(win, ren->centre.x + x, ren->centre.y + y), \
-	ren->color, ren->error));
-	win->set_pixel(win, ren->centre.x - x, ren->centre.y + y, \
-	gamma_average(win->get_pixel(win, ren->centre.x - x, ren->centre.y + y), \
-	ren->color, ren->error));
-	win->set_pixel(win, ren->centre.x + x, ren->centre.y - y, \
-	gamma_average(win->get_pixel(win, ren->centre.x + x, ren->centre.y - y), \
-	ren->color, ren->error));
-	win->set_pixel(win, ren->centre.x - x, ren->centre.y - y, \
-	gamma_average(win->get_pixel(win, ren->centre.x - x, ren->centre.y - y), \
-	ren->color, ren->error));
-}
-
-void	setpixel_inner(t_win *win, int x, int y, t_render_circ *ren)
-{
-	t_compass	*comp;
-	t_hori_line	draw;
-	int			index;
-
-	place_circle_aa(win, x, y, ren);
-	draw.color = ren->color;
-	if (!ren->with_line)
-		return ;
-	comp = ren->comp;
-	draw.y = ren->centre.y + y;
-	if (ren->centre.y + y > ren->c_min_max[MM_MAX_Y])
-	{
-		draw.min_x = ren->centre.x - x;
-		draw.max_x = ren->centre.x + x;
-		draw_horizontal_line(win, &draw);
-	}
-	else
-	{
-		index = ren->centre.y + y - comp->inner.centre.y + comp->inner.radius;
-		draw.min_x = ren->centre.x - x;
-		draw.max_x = comp->circle_x_lim[index].min + comp->inner.centre.x;
-		draw_horizontal_line(win, &draw);
-		if (comp->blur_on == true)
-		{
-			draw.min_x = comp->circle_x_lim[index].min + comp->inner.centre.x;
-			draw.max_x = comp->circle_x_lim[index].max + comp->inner.centre.x;
-			drop_the_blur(win, comp, &draw);
-		}
-		draw.min_x = comp->circle_x_lim[index].max + comp->inner.centre.x;
-		draw.max_x = ren->centre.x + x;
-		draw_horizontal_line(win, &draw);
-	}
-	draw.y = ren->centre.y - y;
-	if (ren->centre.y - y < ren->c_min_max[MM_MIN_Y])
-	{
-		draw.min_x = ren->centre.x - x;
-		draw.max_x = ren->centre.x + x;
-		draw_horizontal_line(win, &draw);
-	}
-	else if (y)
-	{
-		index = ren->centre.y - y - comp->inner.centre.y + comp->inner.radius;
-		draw.min_x = ren->centre.x - x;
-		draw.max_x = comp->circle_x_lim[index].min + comp->inner.centre.x;
-		draw_horizontal_line(win, &draw);
-		if (comp->blur_on == true)
-		{
-			draw.min_x = comp->circle_x_lim[index].min + comp->inner.centre.x;
-			draw.max_x = comp->circle_x_lim[index].max + comp->inner.centre.x;
-			drop_the_blur(win, comp, &draw);
-		}
-		draw.min_x = comp->circle_x_lim[index].max + comp->inner.centre.x;
-		draw.max_x = ren->centre.x + x;
-		draw_horizontal_line(win, &draw);
-	}
-}
+void	setpixel_inner(t_win *win, int x, int y, t_render_circ *ren);
 
 static inline void	setup_ring_draw(t_compass *comp, t_render_circ *ren)
 {
